src/utils.py
import json
import os

def read_json(file_path):
    """
    Reads a JSON file and returns a list of dictionaries with financial transactions.

    :param file_path: Path to the JSON file
    :return: List of dictionaries with financial transactions
    """
    if not os.path.exists(file_path):
        return []

    try:
        with open(file_path, 'r') as file:
            data = json.load(file)
            if not isinstance(data, list):
                return []
            return data
    except (json.JSONDecodeError, IOError):
        return []
